pyproject.toml
[project]
name = "pubmed-affiliation-checker"
version = "0.1.0"
description = "A CLI tool to extract PubMed papers with pharmaceutical/biotech affiliations using LLM"
authors = [
    {name = "Shreenidhi", email = "shreenidhi1903@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "click>=8.1.3",
    "biopython>=1.81",
    "pandas>=2.0.0",
    "openai>=1.0.0",
    "requests >=2.31",
    "python-dotenv>=1.0.0"
]

[tool.poetry]
packages = [{include = "pubmed_affiliation_checker", from = "src"}]

[tool.poetry.scripts]
get-papers-list = "pubmed_affiliation_checker.cli:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
